Refactor Redirect component to separate file
